Add path detection based on libbu location
Up until now we've been using the executable runtime path to deduce the
BRL-CAD root path, but that only works for our own executables.  If an
external program in a different location needs this location, its
executable path will be incorrect.

When the whereami code supports it, prefer constructing the root path
based on libbu's location.  We'll still fall back on trying the
executable if that should fail, but when using libbu's location works it
will be viable even for non-BRL-CAD programs.