Introduce NORETURN as a new function annotation.
Similar to DEPRECATED, this utilizes config testing, but does not require it as the fallback condition is a no-op.  This label marks a function that nominally doesn't return under normal behavior (e.g., exceptions, jumps, and such notwithstanding).  This helps with warnings generated by compilers and code anlaysis by static analyzers.