define an empty setmode() stub.
quelling a windows warning exposed a limitation of #defining setmode() away.  it resulted in a stray (void) cast syntax error, so we instead provide a no-op function instead.  may need another approach if this isn't robust as system setmode() is being masked and we're relying on the compiler to resolve this properly as a static.