need automatic tops reporting and loading when there's just one sensible default.
having to call mged on every .g just to discover what's in there is a usability impacting limiation. the comamnds can/should report the tops available when one isn't specified.  in the absense of something specified, this should greatly reduce the confusion over why nothing is displayed or why objects must be specified.