update name generation for layers
previous method deleted and re-added a layer with the better name but this effected layer indexing.
now, the name of the layer inside the model is unchanged (leaving index untouched) and we check the name
before writing the comb to the db. Geometry name checking is restructured to untangle layer checks.