need to ensure the stack size is unexpectedly large in order for env2c's regex parsing.
unknown portability and just a short-term fix.  still needs to become a proper flag test as this is probably no bueno on windows.
code is also not good if it's ending up with input data on the stack.  it should be reworked dynamic as this is otherwise problematic for other reasons too (e.g., security).