Start setting up for a custom QItemSelectionModel
Trying to work with GED selection sets and the tree view, I'm thinking
more and more that the original intent of the QgSelectionProxyModel
would be better served by the item selection model.  I'm having
difficulty trying to sync the selections with GED, which suggests I'm
trying to do so at the wrong point in the Qt logic flows.

As a preliminary stage, move everything not directly related to the tree
view item highlighting down to QgModel (this doesn't involve any
knowledge of tree view types in QgModel, so we're good there.)  The next
stages will be more involved, since we'll need to change a fair bit of
qged logic to remove the proxy model in favor of QgModel.  The state
management for the edit modes vs. view modes highlighting will also have
to be significantly reworked to function in its new location.  Expect
some disruptions as this proceeds.

End state goal is to have the tree view mode highlighting accurately
reflect the GED selected paths state arrived at by set additions and
removals (and ultimately update the displayed Qt tree selections in
response to the GED select command on the command line, although that'll
be another level of challenge.)  Once all that is in sync, we can return
to the primitive editing question, since we'll have proper selection
sets on which to operate.  For single selections this was never a
challenge, but it is eminently plausible that users will want to select
and (say) move multiple instances of multiple combs simultaneously, and
we don't want our infrastructure to come up short due to design
defects...