Use the selection model to restrict selections
For .g tree widget purposes, we don't want to simultaneously have
selected a child and a parent along the same tree path - conceptually,
it is difficult to know what that implies for operations like editing
and drawing (if tire.r and rim.c are both selected, do we only want
tire.r?  Or only tire.r/rim.c? - in that case, if tire.r is selected,
rim.c should be cleared.  Conversely, if tire.r/rim.c is selected, the
parent tire.r selection is cleared so we can focus on rim.c)

However, we DO allow multiple selections so long as each selection is on
its own independent db_fullpath in the tree - one selection per path.
So while a parent comb may have multiple children selected
simultaneously, if the parent comb is selected all of the child
selections are cleared, since the parent comb is in all of those paths.

We still need to get this syncing with the GED sets, but that will
require a slight modification to what we prototyped earlier for those
sets.  We don't want to automatically collapse the selected paths to
their most compact form - if we have a comb instance T/C with children
1, 2 and 3 and we select T/C/1 T/C/2 and T/C/3, the implication is we
want to edit those three instance matrices rather than the T/C instance
matrix.  However, since the 1,2,3 selections would constitute a full
selection of all of C's children, the collapse operation in GED would
shrink the path to T/C.