loudly protest that musl captures our librt references at runtime as erroneous behavior
and provide an automatic workaround.  nobody should rely on this workaround, and it's not portable (assumes x86_64).  a better test would be to run the /lib/ld-musl-x86_64.so.1 dynamic linker and see if it still exhibits the capture behavior, or to specify LD_PRELOAD as environment on the tools that must run during compilation.