Checkpoint test of a different selection approach
Experimenting with a large comb tree suggests that Qt's QTreeView
scrolling performance drops off when a large number of rows are
selected.  See if we can use the mechanism we already have in place for
highlighting drawing to do our own "selection" which doesn't use the Qt
mechanism.  Preliminary testing suggests this will be more performant,
so stashing the current state.