Adjust how scene obj settings are propagated
When selecting an LoD mesh, the highlighting wasn't getting propagated
down to the view specific objects.  By default, we don't want the
non-LoD settings to be independent on these objects, so adjust the scene
object to optionally allow specifying a pointer to another object's
settings to allow for simple sharing.

We also need to let the drawing routines know to rebuild their dlists if
the settings change (since normally they're rebuilt only when the data
needs to adjust to something like a new LoD view.)  Add a bu_obj_stale
function to handle the bookkeeping.