Make a stab at incorporating a fix for Mac builds
It looks as if there's an issue with the CMake code for PNG with modern
XCode:
https://github.com/PixarAnimationStudios/USD/pull/1946
https://github.com/glennrp/libpng/pull/403

This commit makes a stab at adapting the version from
https://github.com/libpng/libpng/pull/2

onto our (somewhat simplified) png build.

It looks as if libpng doesn't currently have much in the way of an
active upstream (no new releases in a number of years), so will need to
keep an eye out to see what the eventual solution is.