Remove unused files
As far as I can tell, these files aren't used.  Remove them in
preparation for trying to merge in upstream changes.

Looking at netpbm's recent versions, we may have a problem...  We
removed nstring from netpbm some years back due to license issues.  In
the most recent version, it is still using a GPL license, and is
definitely used in a number of places.  Not immediately clear how
difficult it's going to be to identify and merge in upstream changes
while keeping nstring out of the mix...