Count our recursion depth and bail if too deep
This is cheap, but other roads to avoiding crashing via robustness
improvements are looking to be a lot of work, so let's see if this can
keep us running in the meantime...  throw an exception if we recurse too
deeply calling edge events.  We don't get a mesh if this happens, but
hopefully we also don't crash.  Need to test with the non-scrubbed logic
and real geometry, but does avoid wiping out with one of the scrubbed
test cases.