Set PERMISSIONS on file(COPY) to match CMake
This use of the PERMISSIONS options has a downside - it sets ALL the
file permissions, without preserving any of the pre-existing flags.
It's resulting in executables not being runnable when they go through
this logic path.  Work around it by trying to set all the flags to match
what we would get out of default CMake file creation.

This is an unfortunate limitation of file(COPY), and is further
compounded by what as far as I can tell is an inability in CMake to read
the existing permissions off the file portably. Thus we can't even read
the permissions from the file to make an attempt to craft a permissions
preserving copy command manually.  The only other option I know of would
be to make a CMake function wrapper around execing the various platform
specific commands to do this, and if this works I'd prefer to avoid
having to create and manage that solution...