Correct collapse path bookkeeping
Was using the parent comb as a unique key for grouping, and that's not
right - we can get away with that in some cases, but there's
*definitely* no guarantee. pinewood.g exposed the problem by having
child comb reuse, with different higher level parents.  Collapse was
making the wrong decisions because it wasn't binning the paths
correctly.  Need to use the full parent hash as the key to ensure parent
uniqueness.  However, since we are also comparing against the .g
definition of the comb to determine local completeness within the comb
structure, we need to be able to "decode" the unique key path hash to
the parent comb when the time comes for the ground truth comparison.