Remove HIDDEN define in favor of plain 'static'
In my experience the extra layer of the HIDDEN definition isn't
particularly useful - if we're going to be static, we don't want to also
have to worry about name conflicts if a build setting changes, and if
the "don't inline" case is still useful it appears to come up very
rarely.  Motivation for doing this now is that it looks like OpenMesh
also has a HIDDEN definition that collides with ours...  There are
probably other ways to handle that, but might as well use it as a prod
to do the HIDDEN -> static scrub.