OpenMesh/OpenVDB build logic reorg
OpenMesh and OpenVDB logic both pertain to the bot subcommand - move
their build logic to that level.  Get OpenVDB code updated as far as
trying to compile the remesh.cpp file if the flag is enabled - doesn't
build with newer OpenVDB so some updating will be in order when that
code is revisited, but should at least be able to flip the switch and
try the compile now.  Focus is OpenMesh, but the pattern for OpenVDB is
similar so handling both to keep things cleaner.