use vectors in lieu of sets to retain object ordering
std::sets order insertions, which results in objects being alphabatized after
conversions. Continue to use sets to retain uniqueness but write the
combs using a vector to preserve order.
note that this will show some names being added/deleted when running a
gdiff of the set implementation vs the vector implementation. This is
due to the lookup ordering and renaming within the 'polish_output'
section. When orphaned objects part of nested layers are encountered in this function it will
rename according to the database ordering - since we have now changed the
order additions/deletions are reported although all geometry is
accounted for.