checkpoint - most pieces for a simple subdivision
code is generally set up to do catmullclark and/or loop subdivision,
but breaks the 'bot' subcommand in mged - not sure why. Also,
CatmullClarkT.hh throws syntax errors when compiling - also not sure
why.
returning from mesh to bot loop is not tested and may or may not work but follows OpenMesh's
examples for traversal.