Checkpoint working state of ged values cache
This isn't fully correct yet - like LoD, we will need to update this in
the database update cycle after editing, or the UI will display the
wrong values.  However, this cache does let us load the information
needed for drawing at startup without having to crack the avs containers
from the database objects on disk, which is potentially slow for large
objects/models.