add aircode to the set of codes that lc reports
this is primarily so we can catch aircode mismatches with the -m option, but also useful reporting as aircodes significantly change the behavior of nirt, rt, shotlining, etc.  this change adds aircode to the output reporting (columb 4 after los), includes a sort option, and is included in -m mismatch reporting.