First cut at a libged garbage_collect cmd
Since we need to look at the garbage_collect command anyway, go ahead
and recast it as a libged command so it will be more widely available.
Give the opportunity for mishap with this command it needs another set
of eyes, and probably warrants some test cases as well so we can check
it across all platforms more easily.  Approach taken is to do a low
level transcription of the bu_external data, without unpacking further.