stepcode is now upstream 0.8.1 unmodified.
this represents a pretty huge change from an code and architecture perspective (and hundreds of commits of change), but hopefully not much if any affect on users and applications as indicated by testing.  code is more standards compliant, lots of memory and security issues fixed, some restructuring and renaming, and loads more.