Set better management of rt_wdb I/O stream types
MGED is opening two wdb streams on a single dbip - one inmem, and one
disk (via libged).  In principle this isn't forbidden by the APIs, but
it causes a practical difficulty in that db_i stores only one dbi_wdbp
pointer.  If code expects its assigned wdbp to be present in
dbip->dbi_wdbp and someone else assigns a new wdbp with a different type
in a function call somewhere, Bad (and rather odd) things tend to
happen.  libwdb I/O routines will start working differently than the
caller expected them to, for no apparent reason.

Upon examination, the dbi_wdbp pointer is really supposed to be regarded
as private, and the main thing it controls is how read/write operations
access data. (Separate experimentation in a branch proved that
essentially all the other contents of rt_wdb can be successfully moved
either down to db_i or up to libged/libtclcad/etc. - the type is the key
to libwdb's behavioral changes.)  To handle the MGED scenario of
multiple different simultaneous wdb stream types, we alter the logic to
have db_open and db_close set up the four potentially needed stream
types once, and manage them separately using different private pointers
(this whole "private db_instance vars" setup needs to be moved to a
hidden implementation struct, but that's for later...)

This approach results in wdb_dbopen becoming a pass-through returning
one of the four db_instance managed streams, which at first glance looks
like it might be removing a potential ability of client codes to create
and manage their own rt_wdb structures.   However, the need to keep the
dbip and dbi_wdbp pointers in sync meant that was never a realistic
possibility.  MGED's wdb_obj.c code referencing the rtg_headwdb global
stores rt_wdb pointers in a list, but as near as I can tell this logic
will always result in new rt_wdb pointers associated with new db_open
calls, rather than expecting a new rt_wdb to reference the same dbip.

The new approach also has the advantage of making each use of the wdbp
clear as far as intent - rather than hoping dbi_wdbp is holding the
right stream type for the job, the library or app requests exactly what
it needs explicitly and locally.  wdb_close is currently a no-op -
db_close will do the cleanup - but it is added in most cases anyway to
express the intent of treating the wdb stream interface as a local
phenomena rather than setting up global, db_i wide state.

A consequence of this refactor was to expose a variety of direct uses of
the "private" dbi_wdbp pointer, which are now replaced with calls to
wdb_dbopen.  The next step is to look at libged's ged_wdbp - using this
approach, it should be possible to dispense with it entirely in favor of
dbip + wdb_dbopen.

I note in passing that at the moment rt_wdb is being used to track
counting and prefixing strings - as near as I can tell this logic is
used only in libged commands, not in librt or libwdb.  We should
consider whether those variables would make more sense living in libged
as well.  Tolerance and resource variables, on the other hand, almost
certainly belong with the dbip itself - in particular, moving resource
from rt_wdb to db_instance will almost certainly be a major step towards
removal of rt_uniresource as a global, and thanks to the rt_wdb struct
holding a dbip pointer may even be minimally impacting.  (None of those
changes are in this commit, however.)