Demonstrate scene render diffing using swrast
One of the potentials of the swrast display manager backend is being
able to draw and capture graphical scenes as drawn by the main OpenGL
logic for comparison purposes.  This is a proof-of-principle
demonstration of such a check.

At the moment it is only checking a basic wireframe - we will need to
figure out to to handle more sophisticated features, but some of them
will require significant support (the ert embedded raytracing command,
in particular, needs IPC to move data from rt to the framebuffer.)