First steps toward specifying refresh paths
This is harder than it first appears if we allow full path specifiers -
it becomes the same problem the draw command has to solve of sorting out
full paths.  For now, just check if the leaf node is present in any of
the paths and refresh if it is - that's too aggressive in that it won't
confine the refresh to just the parents of a full path.  If there's a
need we can circle around and figure out something else - probably
populating some of the containers for redraw - but this at least narrows
the scope somewhat.