Comment out stub for brep weld subcommand
This is still a feature we want - it is intended to help in cases where
we end up with conversions that have individual surfaces which are
physically located in space in such a way that they together define a
closed volume, but lack the topological connections that constitute a
proper BRep definition.  However, it's implementation is somewhat
non-trivial and it's probably going to end up unimplemented until we hit
a problem that requires it.