Update man pages for brep command/subcommands
The 7.32.6 release does have the current form of the BRep commands (the
work to reorganize them is considerably older), but it looks like the
man pages weren't adjusted or clarified much when that work was done.
In part that was due to most of brep being a very low level command, not
used much in day-to-day workflows. However, there are a few items like
flip and valid that are of more frequent use to modelers, and when we do
need things like plotting we really need them...  Go ahead and do some
cleanup on the documentation.