Add -N option for specifying overlay view obj name
This is redundant right now with the second argv option to overlay, but
it is setting the stage for the deprecation of that argv behavior in
order to allow overlay to accept multiple plot files as inputs.  -N
allows the original behavior to be achieved by supporting custom view
object naming, once the current behavior is removed.