Make a stab at enabling the librt datum unit test
Originally this test was disabled, because the bu_temp_file trick it
uses didn't work reliably cross platform - in some cases fclose on the
FILE pointer immediately deletes the temp file, with no opportunity for
the program to re-open it with a db_open.  So instead, we provide a db
initialization routine that accepts the FILE pointer and does all
necessary work without ever closing it.  Works on Ubuntu Linux, but
needs cross platform testing.

If this ends up working well, there are more places in the code where we
should be using it - really, any time we need to create a temporary .g
file.  The garbage_collect command is one obvious possibility, but there
are very likely others...