wdb_close needs to db_close for wdb_fopen
The wdb_fopen and wdb_fopen_v functions do an all-up db_create, which
means conceptually they are managing the dbip themselves and need
wdb_close to also take down the dbip.  We don't want that when we're
doing db_open based dbip work, so we let wdb_close know via a flag that
it is responsible for doing db_close when the wdb API is assuming
ownership of the dbip.

Josh noticed this manifesting when a .g created with a GED keep command
produced a .g file that couldn't be moved until MGED was closed (which
released the file.)