Replace wdb_close altogether with db_close
With the new setup, the meat of wdb_close is now just db_close.  Leaving
a wdb_close in, but replace all current uses in the code with the direct
db_close call.  Clarifies intent and simplifies.

Looking at 7.32.6 uses of wdb_close, I think I've got all the spurious
calls introduced during the wdb rework removed - we only want the close
operation when we're intending to close the dbip itself.