Try an alternative approach to QtCADQuad spacers
QSpacerItem doesn't allow us to change color, so we can't use a color
difference to indicate which quadrant is active.  The previous method
used a resizing of the active window's margins, which had the
unfortunate side effect of clearing an embedded framebuffer (if active)
since the window size change invalidated any previously rendered
contents.