Update log with commits since 2020-09-14
This is too aggressive in that it's probably going to incorporate many
commits from earlier patch releases that were cherrypicked, but Git's
approach to storing history makes it very difficult to identify the set
of commits uniquely present in this release but not patch releases.  So,
at least this time, we generate an aggressive ChangeLog.