Fix a problem with v4 ars import
The ars_rd_curve function decrements npts and has a for loop that
continues as long as npts is positive.  When this type is size_t
(converted in 0caf2092a24dd) npts overflows and wipes out.  Instead
of size_t, use the signed ssize_t version so it can go negative.