only check if stdin is a tty.
just because stdout has been redirected doesn't necessarily mean we don't want a gui or don't want to be able to type commands.  it could be that mged's output is being redirected to a file for logging purposes, invoked via some wrapper, but we can still issue commands via stdin.