Replace install(CODE) approach to path correction
Hit an issue with CPack and RPM package generation.  TGZ and TBZ2
versions of this work, but something about the internals of how RPM
package building works with CPack results in CMAKE_INSTALL_PREFIX being
set incorrectly for install(CODE).  Issue has been filed at
https://gitlab.kitware.com/cmake/cmake/-/issues/24212

Until either we figure out a way to make this work with some currently
unknown setting or upstream makes an adjustment to how CPack+RPM handle
this case, fall back on making two copies of the file in question and
using the _install version for the installation process.