Be a little more aggressive about path replacement
Looking at tclConfig.sh and friends, the paths produced by a straight-up
replacement of the build directory with the install directory aren't
enough in all cases to produce valid results.  For the situations where
a Tcl source or build directory are referenced there's not much we can
do beyond pointing at the install directory - the original directories
won't be available in an install - but in some other cases a more
aggressive updating will improve the situation.