install(CODE) needs $ENV{DESTDIR} for CPack RPM
Thanks to Brad King for spotting the missing element - we need to mimic
the cmake_install.cmake scripts and prepend with DESTDIR for RPM:
https://gitlab.kitware.com/cmake/cmake/-/issues/24212#note_1283057

TGZ and TBZ2 don't use DESTDIR, which is why they work without it.
Rather an unintuitive failure mode, but I suppose the use case is also
rather obscure.