More wrapper removal.
Chipping away at simplification and cleanup.  Latest GCC on Ubuntu is
resulting in rendering of shaded mode in QGED with lines at the edges of
triangles when using swrast.  Clang-15, on the same machine, does not.
That suggest something subtle happening in the rendering math near the
edges... urk.  First instinct is to simplify and try to move towards
more standard functions whenever possible to target stable behavior,
although that's admittedly unlikely to magically fix this particular
case...