reorder nirt ray firing order when in libged
when run with graphics, restructure to fire nirt rays in order of
console printing ray -> partition report ray -> overlap report ray
instead of having the console printing ray last. Doing it this way
allows the formatting string to be changed from most to least
restrictive. Allowing for more user options (-f and -A) to be
used.