wip - nested idefs were being handled incorrectly
a model with idefs which use other idefs uncovered a fundamental issue
with the approach for unpacking instance definitions. Go back to the
original method of importing all idefs outside of layers. However,
since names cannot be easily guaranteed unique this way, use the uuid to order
hierarchy for the time being.