add error checking when searching int attributes
when using atol to convert a string, there is no guarantee
that there is an int at all - all strings without an int return 0. So if
searching for an attribute with value = 0 there are potentially
incorrect results.
Additionally, add a check to ensure the converted value fits inside the
long container.