Update NETPBM to 10.73.42
A few notes about this update:  previously we had to do our own removal
of nstring's GPL aspects, replacing it with our own implementation using
C++ strings.  Upstream, although they have not yet propagated it to
their tarball releases, addressed this a different way in their SVN
repository to resolve the GPL issue:

https://sourceforge.net/p/netpbm/code/4419/

Due to the way NETPBM does releases, it will be a long while before this
change is part of the tarball release.  Fortunately, testing indicates
that it can be backported with very minimal changes to the "superstable"
tarball version.

Accordingly, this time we have made a bit more of an effort to keep the
bundled NETPBM looking at least a bit more like the upstream repository.
It is still trimmed down extensively and uses a CMake build rather than
the build logic from upstream, but the directory structure more closely
follows what is present upstream.  We're not using symlinks for the
header locations, but fortunately that adjustment is not difficult to
make and we have had to make very few edits to the C sources directly -
mainly a couple adjustments for nstring and removing code using shhopt
(which we're opting not to keep in this copy of the sources.)

It's still not a vanilla drop-in, but with any luck future updates will
now be less work than the all-up merge that would have been needed by
the previous setup.