Remove embedded copy of bullet, set up to detect
This feature is known not to be in a production ready state, and has not
been tested in quite a while.  The bullet code properly belongs in
src/other/ext but was included here originally while trying to get the
command to work cross platform (particularly on Windows, IIRC.)

Rather than keeping a stale copy of Bullet in the tree for a currently
disabled and non-under-active-development feature, we remove the copy
and set up system detection to allow (in principle) an installed Bullet
to support testing and development of the simulate command.  IFF we get
the feature to a production ready point, we can look at what a proper
src/other/ext integration of Bullet will take.  (Given the peculiarities
of Bullet integration we will probably need the src/other/ext copy most
of the time, but unless/until there's a real need we'll postpone diving
into that particular thorn pile...)

This commit gets us as far as being able to use find_package(Bullet) to
locate a specifically compiled system version of Bullet's static libs,
and is able to to generate a simulate command shared object that will
load on Ubuntu.  simulate-test will compile and start to run - it is not
known currently if it will eventually complete successfully (it will
execute, but the backtrace shows it in bu_semaphore_acquire below an
rt_ct_get call triggered originally by a get_aabb raytrace and there
isn't much CPU activity...)

In addition to whatever is causing the test not to proceed, we probably
need to check and see whether our use of Bullet needs to be modernized.
My recollection is that we also need to do some optimization of
interactions between small and large objects - dropping tank_car on a
large GDAL terrain import, for example, I don't believe ever worked due
to the huge difference in bounding box scales.  We probably need to
project the oriented bounding box of the smaller object through the
larger and construct "interaction" bounding volumes to better define the
context for Bullet.