Make non-mixed mode drawing the default
Per user feedback, the preferred default behavior when specifying a draw
operation is to remove already drawn modes that aren't the specified
mode.  For the old BRL-CAD drawing stack this was the only option - the
new code can accommodate multiple modes for one object, so to both keep
the existing default behavior and allow that mixing if desired we add
the -A option.  Update tests to check both expected default behavior and
functioning of -A option.