Reapply "Reorganize build tools in hierarchy"
This reverts commit 697e70c727b0703f31ebea586a6050c7b8c183f8.

Ah - we were building misc/tools after src/other, which (given the way
we're going to structure this) doesn't work.

If anything in misc/tools needs a library in src/other we're going to
have a problem, but at least right now I don't think that's the case -
libxml is being built without ZLIB support since we don't need it for
Docbook processing.