Correct ordering of misc/tools and src/other build
Ah - PATCHELF_EXECUTABLE wasn't set when we got to src/other/ext when we
moved it to misc/tools, since we weren't building misc/tools first.  As
it happens, right now we can do misc/tools first since no code in that
directory is using src/other libraries.

If we ever get into the situation where a tool in misc/tools requires a
src/other library, we're going to have to reconsider how we're
organizing misc/tools and src/other (or at least, how we maintain the
"don't install" vs "to be installed" distinction), but for now the
reordering should work.