List the generated .h file explicitly
We're still getting occasional compile problems with libexpress that
appear to be an attempt to compile an incompletely generated input.

It looks like the stepcode versions of these files are different than
the ones in BRL-CAD proper - probably would make sense to check which
ones are the newer versions and sync up.