Remove attempt at automatic qged swrast fallback
Hitting another case were OpenGL wasn't working, but this time the main
window (which we found out will also try to use OpenGL if it can)
crashes outright.  However, launching with the '-s' option, which avoids
trying to set up the Qt OpenGL widget in the first place, produces a
running GUI.

Given the "fallback" is problematic, there's little justification for
trying to get the "autodetection" mechanism working reliably (even
assuming that can be done at all with Qt) when the user can just
relaunch qged with -s successfully.  Just print a note letting the user
know what the right option to use is for fallback and exit instead.