Separate active view set concept from the libged view cleanup
We want to be able to add/remove views in a set without GED
automatically taking memory ownership of the bview.  However, if GED is
creating the view, the app won't know about cleaning it up - or the app
may want GED to do the cleanup.  Allow for both possibilities.