Explore mechanisms to allow GED to handle app-owned objects
There's currently a problem with the GED zap command and views
containing objects managed by applications.  bv_clear applies a
bv_obj_put to the scene objects (i.e. destroying them) on the assumption
that the objects are owned by GED.  However, if that's NOT the case,
then we're ripping away objects that the parent code is assuming it
still has.

However, if we mix application owned objects and GED created objects
(such as those from the draw command) in the same scene, we need a way
to be able to zap without disturbing the application level objects.

First thought is to add a way in the scene object to indicate
application level ownership and control, and adjust the relevant
routines to respect that ownership model if it is set.  Not sure if it
will work yet... this should be considered an experiment.  It may be
that the correct thing to do is force applications to not assume that
scene objects will remain valid, but that has other implications - in
particular, a parent application wanting to work with scene objects
would have to have some awareness of when GED has pulled the rug out.