Debugging...
Looks like there is a gremlin in the bounds calculating code...  The
failure of the select LoD test on Windows appears to be driven by some
objects just not being drawn.  Investigating, the csg_wireframe_update
bg_sat_aabb_obb test is failing (resulting in an empty vlist when the
LoD update isn't triggered for that primitive), but it shouldn't be out
of view... Still don't know why it fails on Windows only, but it looks
like on Linux bmin and bmax are inf/-inf while on Windows there are
numerical values.  Smells like something may be accidentally working...
Need to start looking at the bounding logic and when/how it's called,
so add a debug message we can trigger breaking on.