Use the default GED view
This is starting to make sense now.  LIBGED is setting up a default
view, and we were also explicitly creating and adding a view.  That's
fine in and of itself, but we also had an implicit assumption that the
default view would also be the first/only view found when processing.
We didn't actually need to explicitly add the view, but did so anyway
and then subsequently ignored it without realizing it.  Because the C++
container ordering on Linux happened to work out we didn't seen an
issue.  However, on Windows the ordering looks like it ended up
different, resulting in the adaptive CSG plotting setting being applied
to the wrong view.  We then ended up targeting the two different views
for different parts of the processing, with bizarre results.

Fix is to just use the provided GED view and rename it.  Test should now
pass on Windows.