First stages of polygon QEventFilter refactor
The QGED plugin implementation of polygon editing and modification
encodes a lot of Qt mouse binding specific logic for driving polygon
updates.  Ideally we want that mouse interaction logic to be reusable,
should other applications want to manipulate polygons.  However, as
implemented, the event filters are methods on the plugin classes and are
thus specific to those widgets.

However, Qt itself has install/remove mechanisms for filters (which we
are wrapping with QtCADView methods) and doesn't appear to require that
a widget's filter be a method on the class.  If that's the case, it may
offer a possibility - turn the currently class-specific filters into
generic filter objects.  The necessary data pointers can then be defined
generically on the filter QObject and set by the application creating
the filter - the internal filter logic will then just reference the
local variables, with no awareness of the original data structure names.

Lots of questions to determine if this really can work, but if so it
offers what may prove to be a clean way to encapsulate the view polygon
Qt editing logic in a reusable way.