Shift gv_data_vZ out of top level bview struct
It doesn't feel right for the vZ value to be a property of the view - it
makes more sense associated with the polygon objects that use it.  For
the Tcl version of the code just move the var into the Tcl specific
structure, and for the Qt/libbv version put it in the bv_polygon
structure directly.