See if we can add a multi-config build check
For Windows, the Ninja build is fast and efficient but doesn't really
test the Debug/Release multi-config logic the way Visual Studio users
encounter it.  See if we can do a multi-config build with just MGED and
its dependencies to try and exercise that aspect of the build logic on a
regular basis.

Immediate motivation is a re-emergence of the "different names for debug
and release libraries" causing a problem with a Debug Visual Studio
build.  Problem was observed in draw_rework branch, but most likely
cause is build changes due to reverting back to an almost completely
vanilla libpng.  The problem isn't fixed yet, so the expected result for
this is that the Debug build fails when the libpng naming triggers a
copy failure.  If not we'll have to see if there's another way to
trigger the issue...