Start working on generalizing snap logic
Original implementation targeted just the libtclcad data structures.
What we actually want/need here is a more general capability with the
flexibility to select subsets of the view's contents for snapping.

Shouldn't be any behavior changes in this commit - just beginning to add
the necessary containers and flags to allow for this flexibility.