Stash experimental code testing InteractiveAndRobustMeshBooleans
The InteractiveAndRobustMeshBooleans research project at
https://github.com/gcherchi/InteractiveAndRobustMeshBooleans seeks to
implement fast and robust boolean operations on triangle meshes.  The
changes being merged here use the 'libirmb' branch in the fork located
at https://github.com/starseeker/InteractiveAndRobustMeshBooleans to try
this approach with a BRL-CAD CSG tree.

This logic, combined with the libirmb branch, works to trigger the
boolean attempt.  However, it currently fails almost immediately on any
non-trivial CSG input due to how the upstream code is generating output
when coplanar faces (commonly found in CSG models) are involved.  They
don't currently generate manifold outputs in those cases
(https://github.com/gcherchi/InteractiveAndRobustMeshBooleans/issues/13)
which means (for example) a coplanar arb subtraction will likely result
in an open mesh, which is in turn invalid for subsequent processing.

I'm mostly stashing this into main in the hopes that we might be
able to interest someone wanting to do follow-on IRMB work for the "need
manifold output" use-case. In theory, it might be possible to use the
available info in the boolean to assemble a manifold output in those
cases.  The idea would be to extending the original IRMB code to "fill
in" the currently produced non-manifold areas resulting from coplanar
interactions, which should be available (in principle) if the code is
able to "bite out" the areas producing the non-manifold outputs in the
first place.  Doing so would require a fairly deep understanding of the
IRMB processing code to identify the necessary information for coplanar
"patching" and injecting the results output mesh assembly logic.

There appears to be work going on by the upstream team to improve
general robustness with the IRMB code, so we don't want to get too far
away from it - in an ideal world we could implement an "always produce
manifold output" mode that could be contributed back upstream to the
main IRMB codebase.